Mocha tests and use 'send'
