+++ lib/staticGzip.js
@@ -10,7 +10,8 @@
 
 // Commented out as I think that connect is avalible from within express...
 // try {
-    var staticMiddleware = require('connect').static;
+    // var staticMiddleware = require('connect').static;
+
 // } catch (e) {
 //  staticMiddleware = require('express').static;
 // }
@@ -23,8 +24,9 @@ var fs = require('fs'),
     StoreStream = require('./storeStream'),
     FileAsset = require('./fileAsset');
 
-var mime = staticMiddleware.mime,
-    staticSend = staticMiddleware.send;
+var mime = require('mime'),
+    send = require('send')
+    ;
 
 /**
  * Strip `Content-*` headers from `res`.
@@ -60,6 +62,23 @@ exports.filter = function(req, res){
 };
 
 /**
+ * Parse the `req` url with memoization.
+ *
+ * @param {ServerRequest} req
+ * @return {Object}
+ * @api private
+ */
+
+var parseUrl = function(req){
+  var parsed = req._parsedUrl;
+  if (parsed && parsed.href == req.url) {
+    return parsed;
+  } else {
+    return req._parsedUrl = parse(req.url);
+  }
+};
+
+/**
  * By default gzip's static's that match the given regular expression /text|javascript|json/
  * and then serves them with Connects static provider, denoted by the given `dirPath`.
  *
@@ -92,7 +111,7 @@ exports = module.exports = function staticGzip(dirPath, options){
 
     var maxAge = options.maxAge || 86400000,
         contentTypeMatch = options.contentTypeMatch || /text|javascript|json/,
-        clientMaxAge = options.clientMaxAge || 0,
+        clientMaxAge = options.clientMaxAge || 604800000,
         prefix = options.prefix || '',
         names = Object.keys(methods),
         compressionOptions = options.compression || {},
@@ -114,7 +133,13 @@ exports = module.exports = function staticGzip(dirPath, options){
             var o = Object.create(options);
             o.path = name;
             o.maxAge = clientMaxAge;
-            staticSend(req, res, next, o);
+            var urlReq = parseUrl(req).pathname;
+            console.log(dirPath + urlReq);
+            send(req, urlReq)
+                .maxage(clientMaxAge || 0)
+                .root(dirPath)
+                .pipe(res)
+                ;
         }
 
         function setHeaders(stat, asset) {
